lost in los santos
## Business Menu
- shopmenu and business separated
- business handles property management, payments etc
## Maps App
- name and banner customizable by player
- ported from taxi interaction with further optimization
- filters available (address, name, type, zone, distance, availability, interior availability)
- remove locations tab from player info / messages menu
- was planning to add quick gps, but seems redundant
# QoL
- savedoutfits addition added
- Search Criteria (Address, Name, Type, Zone, Max Distance) to filter locations
- Re-arranged information menus
	- Removed landlord menu, merged into Player info
	- Gangs, Zones, Locations moved into World Info

+++ Los Santos RED/lsr/World/Places/DynamicPlaces.cs
@@ -220,7 +220,7 @@ public class DynamicPlaces
         int newRegisterCash = RandomItems.GetRandomNumberInt(200, 1200); //3500;
         if(closestLocation != null)
         {
-            newRegisterCash = RandomItems.GetRandomNumberInt(closestLocation.RegisterCashMin, closestLocation.RegisterCashMax);
+            newRegisterCash = RandomItems.GetRandomNumberInt((int) closestLocation.RegisterCashMin, (int) closestLocation.RegisterCashMax);
         }
         CashRegister newVend = new CashRegister(EntrancePos, heading, "Cash Register", "Cash Register", "", obj, newRegisterCash) { OpenTime = 0, CloseTime = 24 };
         newVend.CanInteractWhenWanted = true;
